feature/134 feat: 로그인이 필요한 기능마다 snackbar 알림 처리

+++ front/src/components/post/CommentItem.vue
@@ -24,7 +24,7 @@
 </template>
 
 <script>
-import { LIKE_BUTTON_TYPE } from "@/utils/constants";
+import { ERROR_MESSAGE, LIKE_BUTTON_TYPE } from "@/utils/constants";
 
 export default {
   name: "CommentItem",
@@ -53,6 +53,13 @@ export default {
       );
     },
     async toggleLike() {
+      if (
+        !sessionStorage.getItem("accessToken") ||
+        sessionStorage.getItem("accessToken") === "guest"
+      ) {
+        this.$store.commit("snackbar/SHOW", ERROR_MESSAGE.LOGIN_REQUIRED);
+        return;
+      }
       this.liked
         ? await this.deleteCommentLike()
         : await this.createCommentLike();
